buildctl: remove dump-bolt command
