+++ cmd/buildctl/debug.go
@@ -11,6 +11,5 @@ var debugCommand = cli.Command{
 	Subcommands: []cli.Command{
 		debug.DumpLLBCommand,
 		debug.DumpMetadataCommand,
-		debug.DumpBoltCommand,
 	},
 }
