+++ cmd/buildctl/debug/dumpmetadata.go
@@ -1,9 +1,12 @@
 package debug
 
 import (
+	"errors"
 	"fmt"
 	"path/filepath"
+	"time"
 
+	"github.com/boltdb/bolt"
 	"github.com/urfave/cli"
 )
 
@@ -23,4 +26,37 @@ var DumpMetadataCommand = cli.Command{
 			return fmt.Sprintf("%q: %s", string(k), string(v))
 		})
 	},
+}
+
+func dumpBolt(dbFile string, stringifier func(k, v []byte) string) error {
+	if dbFile == "" {
+		return errors.New("dbfile not specified")
+	}
+	if dbFile == "-" {
+		// user could still specify "/dev/stdin" but unlikely to work
+		return errors.New("stdin unsupported")
+	}
+	db, err := bolt.Open(dbFile, 0400, &bolt.Options{ReadOnly: true, Timeout: 3 * time.Second})
+	if err != nil {
+		return err
+	}
+	defer db.Close()
+	return db.View(func(tx *bolt.Tx) error {
+		// TODO: JSON format?
+		return tx.ForEach(func(name []byte, b *bolt.Bucket) error {
+			return dumpBucket(name, b, "", stringifier)
+		})
+	})
+}
+
+func dumpBucket(name []byte, b *bolt.Bucket, indent string, stringifier func(k, v []byte) string) error {
+	fmt.Printf("%sbucket %q:\n", indent, string(name))
+	childrenIndent := indent + "  "
+	return b.ForEach(func(k, v []byte) error {
+		if bb := b.Bucket(k); bb != nil {
+			return dumpBucket(k, bb, childrenIndent, stringifier)
+		}
+		fmt.Printf("%s%s\n", childrenIndent, stringifier(k, v))
+		return nil
+	})
 }
